docs: add note about ignoring root package.json for example zips
Previously there was only one package.json at the root of the workspace.
With the introduction of an example with a library, the library folder
contains a package.json that was previously being ignored when creating
the example zip file.

This change only ignores the root package.json because it is replaced with a custom one for the example.